Stereotomy integration with choam and sql state mangement
Clean up stuff not needed
