--- model/src/main/java/com/salesforce/apollo/model/NetworkEndpoint.java
@@ -1,22 +1,0 @@
-/*
- * Copyright (c) 2021, salesforce.com, inc.
- * All rights reserved.
- * SPDX-License-Identifier: BSD-3-Clause
- * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-package com.salesforce.apollo.model;
-
-import java.net.InetAddress;
-
-import com.salesforce.apollo.crypto.Digest;
-
-/**
- * @author hal.hildebrand
- *
- */
-public interface NetworkEndpoint {
-
-    InetAddress endpoint();
-
-    Digest protocol();
-}
